requirements.txt
twitchio
discord.py
python-dotenv
watchdog
requests
openai
ollama
asyncio
bs4
beautifulsoup4
flask
psutil
